replace random symbol in an offensive word with smile, list of offensive words taken from google's list of banned words, discard file after reading it, avoid printing the same message multiple times, automatically remove client generated file after program's finished
