+++ censorialchat/io/exception/UnableToCloseException.java
@@ -1,0 +1,7 @@
+package JavaAutomationLab.censorialchat.io.exception;
+
+public class UnableToCloseException extends Exception {
+    public UnableToCloseException(String message) {
+        super(message);
+    }
+}
