+++ censorialchat/io/exception/UnableToReadException.java
@@ -1,0 +1,7 @@
+package JavaAutomationLab.censorialchat.io.exception;
+
+public class UnableToReadException extends Exception{
+    public UnableToReadException(String message) {
+        super(message);
+    }
+}
