+++ censorialchat/io/interfaces/Packable.java
@@ -1,0 +1,4 @@
+package JavaAutomationLab.censorialchat.io.interfaces;
+
+public interface Packable {
+}
