--- chatting/constant/TimeConstant.java
@@ -1,5 +1,0 @@
-package JavaAutomationLab.chatting.constant;
-
-public class TimeConstant {
-    public static final int TIME_TO_DELAY = 10000;
-}
