--- chatting/io/exception/UnableToCloseException.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.chatting.io.exception;
-
-public class UnableToCloseException extends Exception {
-    public UnableToCloseException(String message) {
-        super(message);
-    }
-}
