--- chatting/io/exception/UnableToReadException.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.chatting.io.exception;
-
-public class UnableToReadException extends Exception{
-    public UnableToReadException(String message) {
-        super(message);
-    }
-}
