--- chatting/io/exception/UnableToWriteException.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.chatting.io.exception;
-
-public class UnableToWriteException extends Exception{
-    public UnableToWriteException(String message) {
-        super(message);
-    }
-}
