--- chatting/io/interfaces/Packable.java
@@ -1,4 +1,0 @@
-package JavaAutomationLab.chatting.io.interfaces;
-
-public interface Packable {
-}
