--- chatting/io/interfaces/StreamReader.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.chatting.io.interfaces;
-
-import JavaAutomationLab.chatting.io.exception.UnableToReadException;
-
-public interface StreamReader {
-    Packable read() throws UnableToReadException, UnableToReadException;
-}
