--- chatting/io/interfaces/StreamWriter.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.chatting.io.interfaces;
-
-import JavaAutomationLab.chatting.io.exception.UnableToWriteException;
-
-public interface StreamWriter {
-    void write(String path, Packable pkg) throws UnableToWriteException, UnableToWriteException;
-}
